+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 0.0.3
 
 ### Added
-- Streaming option for the chat feature (experimental)
-- Option to print latest changelog
+- Ask feature. Similar to `chat`, but the program exits automatically after a single response
+- Option to print latest changelog (found in the options menu or via `termai changelog`)
+- Streaming option for the `chat` feature **(experimental)**
 
 ### Changed
 - Moved model selection out of provider settings